requirements.txt
aiofiles==23.1.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
astroid==2.14.1
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.68
botocore==1.29.68
certifi==2022.12.7
charset-normalizer==2.1.1
dill==0.3.6
discord.py==2.1.1
ffmpeg-python==0.2.0
ffmpeg-smart-trim==0.0.3
frozenlist==1.3.3
future==0.18.3
idna==3.4
isort==5.12.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
openai==0.26.5
platformdirs==3.0.0
ptyprocess==0.7.0
pysubs2==1.6.0
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
watchfiles==0.18.1
wrapt==1.14.1
yarl==1.8.2
